+++ src/main/kotlin/org/jetbrains/research/testspark/helpers/psi/PsiHelperFactory.kt
@@ -1,8 +1,9 @@
-package org.jetbrains.research.testspark.helpers.psiHelpers
+package org.jetbrains.research.testspark.helpers.psi
 
 import com.intellij.openapi.diagnostic.Logger
 import com.intellij.psi.PsiFile
 import com.intellij.psi.PsiJavaFile
+import org.jetbrains.research.testspark.helpers.psi.java.JavaPsiHelper
 
 object PsiHelperFactory {
     fun getPsiHelper(psiFile: PsiFile?): PsiHelper? {
